+++ tests/code/types.test.ts
@@ -3,7 +3,7 @@ import '@/constructor.js';
 import '@/dictionary.js';
 import '@/emptyObject.js';
 import '@/event/types.js';
-import '@/likes.js';
+import '@/shape.js';
 import '@/mapping.js';
 import { describe, expect, test } from 'vitest';
